pyproject.toml
[project]
name = "senior-data-engineer-technical-test"
version = "0.1.0"
description = "Senior DE technical test"
authors = [
    { name = "Jitse-Jan", email = "jitsejan@protonmail.com" }
]
dependencies = [
    "dagster>=1.10.7",
    "dagster-webserver>=1.10.7",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "fabric>=3.2.2",
    "dagster-dbt==0.25.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/senior_data_engineer_technical_test"]
